// workspaces/apps/src/src/app/models/NaturalPerson.ts
export interface NaturalPerson {
  firstName: string | undefined;
  lastName: string | undefined;
  birthDate: Date | undefined;
  birthPlace: string | undefined;
  address: string | undefined;
  pesel: string | undefined;
  idCardNumber: string | undefined;
  phone: number | undefined;
  email: string | undefined;
  nip: number | undefined;
  regon: number | undefined;
  customerNumber: number | undefined;
}